Backend/requirement.txt
asgiref==3.8.1
attrs==24.2.0
Django==4.2.16
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-yasg==1.21.7
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
packaging==24.1
pillow==10.4.0
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1